% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageContourPlot.r
\name{ageContourPlot}
\alias{ageContourPlot}
\title{This function generates a plot of the age model confidence itervals along with a 2D kernal density estimate.}
\usage{
ageContourPlot(model, ...)
}
\arguments{
\item{model}{Output of the \code{ageModel} function}

\item{...}{Optional arguments to be passed to plot legend}
}
\description{
This function generates a plot of the age model confidence itervals along with a 2D kernal density estimate.
}
\details{
This function generates a plot of the age model confidence itervals along with a 2D kernal density estimate of the posterior distribution of each age in both time and position
}
