% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageDepthPlot.R
\name{ageDepthPlot}
\alias{ageDepthPlot}
\title{Make plots of the age probability density functions vs position for several samples}
\usage{
ageDepthPlot(
  ages,
  ageSds,
  positions,
  ids,
  positionThicknesses = rep(0, length(ages)),
  distTypes = rep("G", length(ages)),
  legend = "color",
  scale = 1,
  ...
)
}
\arguments{
\item{ages}{Vector of ages}

\item{ageSds}{Vector of 1-sigma values for ages. Must be the same length and given in the same order as \code{ages}}

\item{positions}{Vector of stratigraphic positions for ages. Must be the same length and given in the same order as \code{ages}}

\item{ids}{Vector of sample names for each age. All samples with the same ids will be combined into a single age PDF. Must be the same length and given in the same order as \code{ages}}

\item{positionThicknesses}{Vector of stratigraphic uncertanties for each age. Specified as half thicknesses. Must be the same length and given in the same order as \code{ages}}

\item{distTypes}{Vector of distribution types to model each age as. Choices are 'G' for Gaussian, and 'U' uniform. Must be the same length and given in the same order as \code{ages}}

\item{legend}{c('color', 'adjacent', NA) type of legend to be drawn. color draws color coded boxes for each sample, adjacent displays sample names next to each PDF. NA omits the legend.}

\item{scale}{scale Scaling factor for age PDFs}

\item{...}{Optional arguments to be passed to plot legend}
}
\description{
Make plots of the age probability density functions vs position for several samples
}
\details{
The \code{ageDepthPlot} function makes a plot of the age probability density functions vs stratigraphic position for several samples. All ages with the same name will be combined into a single PDF.
}
