% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageExtract.r
\name{ageExtract}
\alias{ageExtract}
\title{Extract the highest density intervals for the dated horizons using an ageModel output}
\usage{
ageExtract(model, probability = 0.95)
}
\arguments{
\item{model}{model Output of the \code{ageModel} function}
}
\value{
HDI = specified probability highest density interval of the posterior distribution for each input horizon
}
\description{
Extract the highest density intervals for the dated horizons using an ageModel output
}
