% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bchron_model.R
\name{bchron_model}
\alias{bchron_model}
\title{Run a modified version of the Bchronology algorithm from the R package "Bchron" (Haslett and Parnell, 2008)}
\usage{
bchron_model(
  ages,
  ageSds,
  positions,
  positionThicknesses = rep(0, length(positions)),
  ids,
  distTypes = rep("G", length(ages)),
  iterations = 10000,
  burn = 2000,
  probability = 0.95,
  predictPositions = seq(min(positions), max(positions), length = 500),
  truncateUp = 0,
  extrapUp = 100,
  truncateDown = 1e+10,
  extrapDown = 100
)
}
\arguments{
\item{ages}{Vector of ages}

\item{ageSds}{Vector of 1-sigma values for ages. Must be the same length and given in the same order as \code{ages}}

\item{positions}{Vector of stratigraphic positions for ages. Must be the same length and given in the same order as \code{ages}. Must be input as distance above base of section.}

\item{positionThicknesses}{Vector of stratigraphic uncertainties for each age. Specified as half thicknesses. Must be the same length and given in the same order as \code{ages}}

\item{ids}{Vector of sample names for each age. All samples with the same ids will be combined into a single age PDF. Must be the same length and given in the same order as \code{ages}}

\item{distTypes}{c('G','U') Vector of distribution types to model each age as. Choices are 'G' for Gaussian, and 'U' uniform. Must be the same length and given in the same order as \code{ages}}

\item{iterations}{Number of MCMC iteration to run for. Defaults to 10000}

\item{burn}{Number of initial iterations to discard. Defaults to 2000}

\item{predictPositions}{Vector of stratigraphic positions to evaluate the model at. Defaults to 500 evenly spaced points from the bottom to top of the section}

\item{truncateUp}{Truncation age for extrapolating above the top of the section. Defaults to 0}

\item{extrapUp}{Number of extrapolations to perform above the top of the section. defaults to 100}

\item{truncateDown}{Truncation age for extrapolation below the bottom of the section. Defaults to 1e10 (a really big number)}

\item{extrapDown}{Number of extrapolations to perform below the bottom of the section. defaults to 100}

\item{prob}{Desired confidence to return. Defaults to 95% HDI}

\item{adapt}{Should the proposal standard deviation for model parameters be determined using the adaptive proposal algorithm of Haario et al., 1998. Defaults to TRUE}

\item{mhSD}{Metropolis-Hastings proposal standard deviation for the age parameters. If adapt = TRUE it will be determined by the model run. Otherwise a vector of standard deviations must be specified with length = \code{unique(ids)}}

\item{psiSD}{Metropolis-Hastings proposal standard deviation for the Compound Poisson-Gamma scale parameter. If adapt = TRUE it will be determined by the model run.}

\item{muSD}{Metropolis-Hastings proposal standard deviation for the Compound Poisson-Gamma mean parameter. If adapt = TRUE it will be determined by the model run.}
}
\value{
HDI = specified probability highest density interval for the model run

model = Raw model predictions for each MCMC iteration

thetas = The posterior values for each dated horizon from each MCMC run. burn is not removed.

predictPositions = Stratigraphic positions where the model was evaluated. Same as \code{predictPositions}

mu, psi = The posterior values for the Compound Poisson-Gamma mean and scale parameters. burn is not removed.

ageGrid = Grid that age PDFs were evaluated over. Useful for plotting

likelihoods = Age PDFs for each set of dated samples. Useful for plotting

nAges = Number of ages that were combined into each sample. Useful for plotting

masterPositions = Stratigraphic positions for each set of combined ages. Useful for plotting

ids = Vector of unique sample names. Useful for plotting
}
\description{
Run a modified version of the Bchronology algorithm from the R package "Bchron" (Haslett and Parnell, 2008)
}
