% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterPlot.r
\name{parameterPlot}
\alias{parameterPlot}
\title{Make plots of model parameters from the output of an ageModel run}
\usage{
parameterPlot(model, PDF = F)
}
\arguments{
\item{model}{Output of the \code{ageModel} function}

\item{PDF}{should plots be printed to the current working directory as PDF files? Defaults to FALSE}
}
\description{
Make plots of model parameters from the output of an ageModel run
}
