% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorPlot.r
\name{posteriorPlot}
\alias{posteriorPlot}
\title{Make plots of modeled posteriors from the output of an ageModel run}
\usage{
posteriorPlot(model, prob = 0.95, sigF = 3)
}
\arguments{
\item{model}{Output of the \code{ageModel} function}

\item{prob}{Desired level of probability for HDI's. Defaults to 95%}

\item{sifF}{number of decimal places to display on histograms}
}
\description{
Make plots of modeled posteriors from the output of an ageModel run
}
