% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxyPlot.r
\name{proxyPlot}
\alias{proxyPlot}
\title{This function plots the result of a time-integrated smoothed proxy record}
\usage{
proxyPlot(proxySmoothOutput, model, stratSmoothParameter = "default")
}
\arguments{
\item{proxySmoothOutput}{the output of the proxySmooth function}

\item{model}{output of the age model used to predict ages and smooth the data}

\item{stratSmoothParameter}{optional parameter for smoothing the proxy data vs stratigraphic position. must be meet the same criteria as \code{smoothParameter} in \code{proxySmooth}.
Default method for spline fitting uses the same \code{smoothParameter} as proxySmooth. Default method for moving average fitting will estimate a smoothing parameter based on the median accumulation
rate and \code{smoothParameter} from \code{proxySmooth}.}
}
\description{
This function plots the result of a time-integrated smoothed proxy record
}
