% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxySmooth.r
\name{proxySmooth}
\alias{proxySmooth}
\title{This function uses a Monte Carlo aproach to repetedly fit a smoothing model to some proxy data incorperating the model uncertanties in age}
\usage{
proxySmooth(proxy, proxySD = rep(0, length(proxy)), agePredictOutput, grid,
  method = c("spline", "movingAverage", "polynomial"), smoothParameter = 1,
  HDI = 0.95)
}
\arguments{
\item{proxy}{Vector proxy compostions. must be numeric}

\item{proxySD}{optional vector of 1 sigma uncertanties for \code{proxy}. Must be the same length as \code{proxy}}

\item{agePredictOutput}{Output from \code{agePredict} for the stratigraphic positions of \code{proxy}}

\item{grid}{vector of positions to evaluate the smoothing model at}

\item{method}{Smoothing method to use. currently supported methods are 'spline', 'movingAverage' and 'polynomial'}

\item{smoothParameter}{parameter to use for smoothing. If method = 'spline' \code{smoothParameter} should be between 0 and 1. see \code{smooth.spline} for details. If method = 'movingAverage' \code{smoothParameter} is the width of the smoothing window in units of time used by the age-depth model. If method = 'polynomial' \code{smoothParameter} is the degree of polynomial to fit}

\item{HDI}{desired probability interval to return. Must be between 0 and 1}
}
\description{
This function uses a Monte Carlo aproach to repetedly fit a smoothing model to some proxy data incorperating the model uncertanties in age
}
