% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlot.r
\name{modelPlot}
\alias{modelPlot}
\title{This function generates two age model plots given the output of an ageModel run.}
\usage{
modelPlot(model, agePredictOutput = NA, predictLabels = T, scale = 1,
  type = "PDF", ...)
}
\arguments{
\item{model}{Output of the \code{ageModel} function}

\item{agePredictOutput}{Output of the agePredict function. If specified, points and error bars will be added to the age model plots for the predicted points}

\item{predictLabels}{should predicted ages for points be displayed?}

\item{scale}{Scaling factor for age PDFs}

\item{type}{c('PDF', Contour') Shound probability be displayed as likelihood input PDFs or as contours of posterior probability. Defaults to PDF}

\item{...}{Optional arguments to be passed to plot (xlim, ylim, xlab, ylab, main)}
}
\description{
This function generates two age model plots given the output of an ageModel run.
}
