% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPlot.r
\name{modelPlot}
\alias{modelPlot}
\title{This function generates two age model plots given the output of an ageModel run.}
\usage{
modelPlot(model, agePredictOutput = NA, predictLabels = "both", scale = 1,
  type = "PDF", legend = "color", ...)
}
\arguments{
\item{model}{Output of the \code{ageModel} function}

\item{agePredictOutput}{Output of the agePredict function. If specified, points and error bars will be added to the age model plots for the predicted points}

\item{predictLabels}{c('ages','ids','both', NA) should predicted ages and names for for points be displayed? Defaults to display both. Set to NA for no labels}

\item{scale}{Scaling factor for age PDFs}

\item{type}{c('PDF', contour') Shound probability be displayed as likelihood input PDFs or as contours of posterior probability. Defaults to PDF}

\item{legend}{c('color', 'adjacent', NA) type of legend to be drawn. color draws color coded boxes for each sample, adjacent displays sample names next to each PDF. NA omits the legend.}

\item{...}{Optional arguments to be passed to plot (xlim, ylim, xlab, ylab, main)}
}
\description{
This function generates two age model plots given the output of an ageModel run.
}
